Fix use of usernameField
Fixes Passport not getting the specified usernameField in the plugin.
This implies a change in how it's used with password, before it was
used like this:
passport.use(new LocalStrategy(User.authenticate()));

Now it's used like this:
passport.use(User.authenticate());

It's simplier and I don't think it's useful getting that object out of
the plugin. The only 'problem' is the require added to the plugin, but
well, this plugin also require passport to be present so it's
affordable.